Refactor: Moved commands of same group to folders
